+++ externals/kyuubi-spark-sql-engine/src/main/scala/org/apache/kyuubi/engine/spark/SparkSQLEngine.scala
@@ -49,6 +49,13 @@ private[spark] final class SparkSQLEngine(name: String, spark: SparkSession)
     }
   }
 
+  override def start(): Unit = {
+    super.start()
+    // Start engine self-terminating checker after all services are ready and it can be reached by
+    // all servers in engine spaces.
+    backendService.sessionManager.startTerminatingChecker()
+  }
+
   override protected def stopServer(): Unit = {
     countDownLatch.countDown()
   }
@@ -60,7 +67,7 @@ object SparkSQLEngine extends Logging {
 
   private val user = Utils.currentUser
 
-  private[spark] val countDownLatch = new CountDownLatch(1)
+  private val countDownLatch = new CountDownLatch(1)
 
   def createSpark(): SparkSession = {
     val sparkConf = new SparkConf()
